+++ practice/tasks/Decorator/Epam.NetMentoring.Decorator.Calculator/IOperation.cs
@@ -4,6 +4,6 @@
     {
         //IT: Result from divide/multiply is not always int
         //IS: good point but want to discuss it before changing 
-        int GetResult();
+        double GetResult();
     }
 }
